Arreglado error en subida

+++ routes/index.js
@@ -19,32 +19,26 @@ router.get('/', function (req, res, next) {
 });
 
 router.post('/uploadfile', async (req, res) => {
-  if (req.file!="") {
-    console.log(req.file);
-    var result = await cloudinary.v2.uploader.upload(req.file.path);
-    console.log(result);
-    await fs.unlink(req.file.path);
-    var image_url = result.url;
-    var public_id = result.public_id;
-    var descripcion = req.body.descripcion;
-    var titulo = req.body.titulo;
-    var subir = { titulo, descripcion, image_url, public_id };
-  } else {
-    res.render('/', {
-      error: true, messaje: 'Todos los campos son Requeridos'
-    });
-  }
   try {
-    if (image_url != "" && public_id != "" && req.body.titulo != "" && req.body.descripcion != "") {
+    if (req.file != undefined && image_url != "" && public_id != "" && req.body.titulo != "" && req.body.descripcion != "") {
+      console.log(req.file);
+      var result = await cloudinary.v2.uploader.upload(req.file.path);
+      console.log(result);
+      await fs.unlink(req.file.path);
+      var image_url = result.url;
+      var public_id = result.public_id;
+      var descripcion = req.body.descripcion;
+      var titulo = req.body.titulo;
+      var subir = { titulo, descripcion, image_url, public_id };
       await imagenModel.insertImagen(subir);
     } else {
-      res.render('/', {
+      res.render('index', {
         error: true, messaje: 'Todos los campos son Requeridos'
       });
     }
   } catch (error) {
     console.log(error);
-    res.render('/', {
+    res.render('index', {
       error: true, messaje: 'No se Cargo La Novedad'
     })
   }
